# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateAddressBookRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateContactRequest:
    properties:
      vcard_data:
        example: BEGIN:VCARD...
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.DeleteAddressBookRequest:
    properties:
      confirmation:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.UpdateAddressBookRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_domain_addressbook.AddressBook:
    properties:
      contacts:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_addressbook.AddressObject'
        type: array
      createdAt:
        type: string
      ctag:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      syncToken:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_user.User'
      userID:
        type: integer
      uuid:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_addressbook.AddressObject:
    properties:
      addressBook:
        $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_addressbook.AddressBook'
      addressBookID:
        type: integer
      contentLength:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Primary email
        type: string
      etag:
        type: string
      familyName:
        type: string
      formattedName:
        description: Denormalized fields for search
        type: string
      givenName:
        type: string
      id:
        type: integer
      organization:
        type: string
      path:
        type: string
      phone:
        description: Primary phone
        type: string
      uid:
        description: vCard UID
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
      vcardData:
        type: string
      vcardVersion:
        description: '"3.0" or "4.0"'
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_domain_contact.Address:
    properties:
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      type:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.Contact:
    properties:
      addressbook_id:
        description: UUID of the AddressBook
        type: string
      addresses:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Address'
        type: array
      birthday:
        description: Other
        type: string
      created_at:
        type: string
      emails:
        description: Lists
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Email'
        type: array
      etag:
        type: string
      family_name:
        type: string
      formatted_name:
        type: string
      given_name:
        type: string
      id:
        description: UUID of the AddressObject
        type: string
      middle_name:
        type: string
      nickname:
        type: string
      notes:
        type: string
      organization:
        description: Organization
        type: string
      phones:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Phone'
        type: array
      photo_url:
        description: Constructed URL for the photo
        type: string
      prefix:
        description: Name components
        type: string
      suffix:
        type: string
      title:
        type: string
      uid:
        description: vCard UID
        type: string
      updated_at:
        type: string
      urls:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.URL'
        type: array
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.Email:
    properties:
      primary:
        type: boolean
      type:
        type: string
      value:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.Phone:
    properties:
      primary:
        type: boolean
      type:
        type: string
      value:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_contact.URL:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_domain_user.OAuthConnection:
    properties:
      accessToken:
        description: encrypted
        type: string
      createdAt:
        type: string
      id:
        type: integer
      provider:
        description: google, microsoft, custom
        type: string
      providerEmail:
        type: string
      providerID:
        description: sub claim from OIDC
        type: string
      refreshToken:
        description: encrypted
        type: string
      tokenExpiry:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_user.User'
      userID:
        type: integer
    type: object
  github_com_jherrma_caldav-server_internal_domain_user.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      displayName:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: integer
      isActive:
        type: boolean
      oauthConnections:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_user.OAuthConnection'
        type: array
      passwordHash:
        type: string
      updatedAt:
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_usecase_contact.ListOutput:
    properties:
      contacts:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_adapter_http.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: error_code
        type: string
      message:
        example: Human readable message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HTTP Basic Authentication for DAV endpoints
  title: CalDAV/CardDAV Server API
  version: "1.0"
paths:
  /addressbooks:
    get:
      description: Get all address books for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              addressbooks:
                items:
                  $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_addressbook.AddressBook'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: List address books
      tags:
      - AddressBooks
    post:
      consumes:
      - application/json
      description: Create a new address book
      parameters:
      - description: Address book details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateAddressBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_addressbook.AddressBook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create address book
      tags:
      - AddressBooks
  /addressbooks/{addressbook_id}/contacts:
    get:
      description: Get contacts from address book
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Limit (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Sort field (default name)
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_usecase_contact.ListOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: List contacts
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      description: Create a new contact
      parameters:
      - description: Address Book ID
        in: path
        name: addressbook_id
        required: true
        type: integer
      - description: Contact details
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_contact.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create contact
      tags:
      - Contacts
...
  /addressbooks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete address book
      parameters:
      - description: Address Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.DeleteAddressBookRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete address book
      tags:
      - AddressBooks
    get:
      description: Get address book by ID
      parameters:
      - description: Address Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_addressbook.AddressBook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Get address book
      tags:
      - AddressBooks
    patch:
      consumes:
      - application/json
      description: Update address book details
      parameters:
      - description: Address Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.UpdateAddressBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_domain_addressbook.AddressBook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update address book
      tags:
      - AddressBooks
  /addressbooks/{id}/contacts:
    post:
      consumes:
      - application/json
      description: Create a new contact in address book
      parameters:
      - description: Address Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contact VCard
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'Ref: domain.AddressObject'
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create contact
      tags:
      - Contacts
  /addressbooks/{id}/export:
    get:
      description: Export address book as vCard
      parameters:
      - description: Address Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/vcard
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Export address book
      tags:
      - Import/Export
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...